Detect Stop Releases

+++ real-upgrade.sh
@@ -15,6 +15,14 @@ PKG_DIR=/usr/share/sfos-moto_msm8960_jbbl-adaptation
 echo -e "\n=== RELEASE: $RELEASE ===\n"
 ssu release $RELEASE
 
+echo -e "\n=== Available space in rootfs: ===\n"
+df -h /
+echo -e "\n=== Do you want to continue? [Y/n] ===\n"
+read yn
+if [ x$yn == x"n" ]; then
+    exit 1
+fi
+
 # disable openrepos
 OPENREPOS="$(ssu lr | sed -n '/Enabled repositories (user)/,/Disabled/p' | awk '/openrepos/{print $2}')"
 echo -e "$OPENREPOS" > $PKG_DIR/.disabled_repos
@@ -27,14 +35,6 @@ patchmanager --unapply-all || true
 ssu lr
 zypper clean -a
 zypper ref -f
-
-echo -e "\n=== Available space in rootfs: ===\n"
-df -h /
-echo -e "\n=== Do you want to continue? [Y/n] ===\n"
-read yn
-if [ x$yn == x"n" ]; then
-    exit 1
-fi
 
 version --dup
 $PKG_DIR/moto_msm8960_jbbl.sh
